Add support for cooperative REPLS
Sometimes when embedding a REPL server in an application with an
event loop, we don't want expressions sent to the REPL to be
evaluated asynchronously, rather we want to be able to control
where they are evaluated in the main thread.

To this end, we introduce a "cooperative" option for the REPL
server, which means that eval ops will put the input on a queue
which will be processed to complete the evaluation only when
`poll-repl-server` is called.